add admit bed list and admit bed list with patient detail

+++ HchApiPlatform/Models/N.cs
@@ -9,13 +9,13 @@ namespace HchApiPlatform.Models
         public string? NsName { get; set; }
         public string? NsType { get; set; }
         public string? UddFlag { get; set; }
-        public byte? UddDeliverTime { get; set; }
+        public int? UddDeliverTime { get; set; }
         public string? DeptCode { get; set; }
         public string? OpenFlag { get; set; }
         public string? UddAutoDcFlag { get; set; }
-        public byte? IdeSpreadStartTime { get; set; }
+        public int? IdeSpreadStartTime { get; set; }
         public string? HospLocation { get; set; }
-        public byte? UddProcessOrder { get; set; }
+        public int? UddProcessOrder { get; set; }
         public string? NsEnglishName { get; set; }
     }
 }
